move neighbor db code to util class
